+++ R/read.R
@@ -1,3 +1,10 @@
+sf_stringsAsFactors = function() {
+	if (getRversion() < "4.1.0")
+		default.stringsAsFactors()
+	else
+		FALSE
+}
+
 set_utf8 = function(x) {
 	n = names(x)
 	Encoding(n) = "UTF-8"
@@ -44,9 +51,9 @@ set_utf8 = function(x) {
 #'   of LineString and MultiLineString, or of Polygon and MultiPolygon, convert
 #'   all to the Multi variety; defaults to \code{TRUE}
 #' @param stringsAsFactors logical; logical: should character vectors be
-#'   converted to factors?  The `factory-fresh' default is \code{TRUE} for
-#'   \code{st_read} and \code{FALSE} for \code{read_sf}, but this can be changed
-#'   globally by e.g. the R command \code{options(stringsAsFactors = FALSE)}.
+#'   converted to factors?  Default for \code{read_sf} or R version >= 4.1.0 is 
+#' \code{FALSE}, for \code{st_read} and R version < 4.1.0 equal to
+#' \code{default.stringsAsFactors()}
 #' @param int64_as_string logical; if TRUE, Int64 attributes are returned as
 #'   string; if FALSE, they are returned as double and a warning is given when
 #'   precision is lost (i.e., values are larger than 2^53).
@@ -134,7 +141,8 @@ st_read.default = function(dsn, layer, ...) {
 }
 
 process_cpl_read_ogr = function(x, quiet = FALSE, ..., check_ring_dir = FALSE,
-		stringsAsFactors = ifelse(as_tibble, FALSE, default.stringsAsFactors()), geometry_column = 1, as_tibble = FALSE) {
+		stringsAsFactors = ifelse(as_tibble, FALSE, sf_stringsAsFactors()), 
+		geometry_column = 1, as_tibble = FALSE) {
 
 	which.geom = which(vapply(x, function(f) inherits(f, "sfc"), TRUE))
 
@@ -199,8 +207,8 @@ process_cpl_read_ogr = function(x, quiet = FALSE, ..., check_ring_dir = FALSE,
 #' to the current working directory (see \link{getwd}). "Shapefiles" consist of several files with the same basename
 #' that reside in the same directory, only one of them having extension \code{.shp}.
 #' @export
-st_read.character = function(dsn, layer, ..., query = NA, options = NULL, quiet = FALSE, geometry_column = 1L, type = 0,
-		promote_to_multi = TRUE, stringsAsFactors = default.stringsAsFactors(),
+st_read.character = function(dsn, layer, ..., query = NA, options = NULL, quiet = FALSE, geometry_column = 1L, 
+		type = 0, promote_to_multi = TRUE, stringsAsFactors = sf_stringsAsFactors(),
 		int64_as_string = FALSE, check_ring_dir = FALSE, fid_column_name = character(0),
 		drivers = character(0), wkt_filter = character(0)) {